requirements.txt
google-cloud-bigquery
google-cloud-aiplatform
google-generativeai
vertexai
tqdm
pandas
matplotlib
seaborn
streamlit
